+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-22T11:24:22.432Z",
+    "2026-01-22T13:36:18.119Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {