+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-01-22T11:24:25.787Z",
+    "2026-01-22T13:36:21.569Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {