+++ packages/orq-rc/RELEASES.md
@@ -6129,3 +6129,13 @@ Based on:
 - [typescript v4.2.0-rc.61] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.61] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.61 - packages/orq-rc
+
+## 2026-01-22 13:41:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.693.0 (2.798.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.63] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.63] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.63 - packages/orq-rc